refactor out duplication
