+++ webview/src/setup/components/experiments/Experiments.tsx
@@ -1,22 +1,19 @@
 import React from 'react'
-import { useDispatch, useSelector } from 'react-redux'
-import { SetupSection } from 'dvc/src/setup/webview/contract'
+import { useSelector } from 'react-redux'
 import { NoData } from './NoData'
 import { NeedsGitCommit } from './NeedsGitCommit'
 import { showExperiments } from '../messages'
 import { EmptyState } from '../../../shared/components/emptyState/EmptyState'
 import { IconButton } from '../../../shared/components/button/IconButton'
 import { Beaker } from '../../../shared/components/icons'
-import { Button } from '../../../shared/components/button/Button'
-import { updateSectionCollapsed } from '../../state/webviewSlice'
 import { SetupState } from '../../store'
+import { FocusDvcSection } from '../shared/FocusDvcSection'
 
 type ExperimentsProps = {
   isDvcSetup: boolean
 }
 
 export const Experiments: React.FC<ExperimentsProps> = ({ isDvcSetup }) => {
-  const dispatch = useDispatch()
   const { needsGitCommit, hasData } = useSelector(
     (state: SetupState) => state.experiments
   )
@@ -26,19 +23,7 @@ export const Experiments: React.FC<ExperimentsProps> = ({ isDvcSetup }) => {
       <EmptyState isFullScreen={false}>
         <h1>DVC is not setup</h1>
         <p>DVC needs to be setup before you can access experiments.</p>
-        <Button
-          onClick={() =>
-            dispatch(
-              updateSectionCollapsed({
-                [SetupSection.DVC]: false,
-                [SetupSection.EXPERIMENTS]: true,
-                [SetupSection.REMOTE]: true,
-                [SetupSection.STUDIO]: true
-              })
-            )
-          }
-          text="Setup DVC"
-        />
+        <FocusDvcSection />
       </EmptyState>
     )
   }
