+++ webview/src/setup/components/remote/DvcUninitialized.tsx
@@ -1,28 +1,10 @@
 import React from 'react'
-import { useDispatch } from 'react-redux'
-import { SetupSection } from 'dvc/src/setup/webview/contract'
 import { EmptyState } from '../../../shared/components/emptyState/EmptyState'
-import { Button } from '../../../shared/components/button/Button'
-import { updateSectionCollapsed } from '../../state/webviewSlice'
+import { FocusDvcSection } from '../shared/FocusDvcSection'
 
-export const DvcUninitialized: React.FC<{}> = () => {
-  const dispatch = useDispatch()
-  return (
-    <EmptyState isFullScreen={false}>
-      <h1>DVC is not initialized</h1>
-      <Button
-        text="Setup DVC"
-        onClick={() =>
-          dispatch(
-            updateSectionCollapsed({
-              [SetupSection.DVC]: false,
-              [SetupSection.EXPERIMENTS]: true,
-              [SetupSection.REMOTE]: true,
-              [SetupSection.STUDIO]: true
-            })
-          )
-        }
-      />
-    </EmptyState>
-  )
-}
+export const DvcUninitialized: React.FC<{}> = () => (
+  <EmptyState isFullScreen={false}>
+    <h1>DVC is not initialized</h1>
+    <FocusDvcSection />
+  </EmptyState>
+)
