+++ webview/src/setup/components/studio/Studio.tsx
@@ -1,9 +1,10 @@
 import React from 'react'
 import { useSelector } from 'react-redux'
+import { STUDIO_URL } from 'dvc/src/setup/webview/contract'
 import { Connect } from './Connect'
 import { Settings } from './Settings'
-import { CliIncompatible } from './CliIncompatible'
 import { SetupState } from '../../store'
+import { CliIncompatible } from '../shared/CliIncompatible'
 
 export const Studio: React.FC<{
   cliCompatible: boolean
@@ -12,7 +13,13 @@ export const Studio: React.FC<{
   const { isStudioConnected } = useSelector((state: SetupState) => state.studio)
 
   if (!cliCompatible) {
-    return <CliIncompatible />
+    return (
+      <CliIncompatible>
+        <p>
+          Locate DVC to connect to <a href={STUDIO_URL}>Studio</a>
+        </p>
+      </CliIncompatible>
+    )
   }
 
   return isStudioConnected ? (
